requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pyotp==2.9.0
qrcode==7.4.2
pillow==10.0.0
pydantic==2.4.0
logging-config==1.0.0
pytest==7.4.2
pytest-cov==4.1.0
black==23.10.0
flake8==6.1.0